Fixed compiler warnings

+++ modules/CheckLogFile/realtime_thread.cpp
@@ -117,7 +117,7 @@ void real_time_thread::thread_proc() {
       dwWaitTime = 0;
     else if (dur)
       dwWaitTime = dur->total_milliseconds();
-    DWORD dwWaitReason = WaitForMultipleObjects(logs.size() + 1, handles, FALSE, dwWaitTime);
+    DWORD dwWaitReason = WaitForMultipleObjects(static_cast<DWORD>(logs.size() + 1), handles, FALSE, dwWaitTime);
     if (dwWaitReason == WAIT_TIMEOUT) {
       // we take care of this below...
     } else if (dwWaitReason == WAIT_OBJECT_0) {
